build: allow splitting packages out (e.g. gcc-libs from gcc)
This is useful for large packages on which many other packages depend, such as
gcc (>800 MB). We can now split out a <70 MB gcc-libs package to reduce disk
space.

Note that the files which are split out must be relocatable, i.e. they must not
depend on the location to which they are installed. This is because the package
is built only once, with the --prefix= configure flag pointing to the main
package’s name, not the split packages name.

An example looks like this:

  split_package: <
    name: "gcc-libs"
    claim_glob: "out/lib64/libgcc_s.so*"
    claim_glob: "out/lib64/libstdc++*"
    claim_glob: "out/lib64/libgomp*"
    claim_glob: "out/lib/gcc/x86_64-pc-linux-gnu/8.2.0/crt*"
  >

This feature should be used sparingly: the default expectation is that the
package name matches DISTRIROOT/pkgs/<name>.