build: prune circular runtime dependencies
This is a problem when bumping packages in the bootstrap set, e.g. gcc-libs.
gcc-libs depends on itself: you need gcc-libs to build gcc-libs.

When building the new version (e.g. 4), the old version (3) will be found when
finding runtime dependencies, and its runtime dependencies are pulled into the
mix when resolving them.