python3: fix -cross=i686 build
This required one patch and hard-coding a few -L flags to guide the linker
because Setup.py doesn’t use pkg-config to find zlib and ncurses on Linux :(

Also, set --enable-ipv6 to avoid an overly pessimistic check for a 19-year old
getaddrinfo bug: https://twitter.com/zekjur/status/1255045485754277888

…and ac_cv_file__dev_{ptmx,ptc} because those can’t be checked when
cross-compiling. Contrary to the --enable-ipv6 flag above, this was easy to
understand because of a very direct message from configure.

Also switch to placing Modules/Setup.local in the build dir (copied from the
source dir, via an extra_file directive) instead of modifying Modules/Setup in
the build steps or curses.patch. Let’s see if the ease of use of the separate
file is worth it over time, in terms of maintenance effort.